### 2020-06-04 #382 @bretsw
tidytags: Simple Collection and Powerful Analysis of Twitter Data
Submitting Author Name: Bret Staudt Willet
Submitting Author Github Handle: @bretsw
Repository: https://github.com/bretsw/tidytags
Version submitted:   0.1.0
Submission type: Standard
Due date for @llrs: 2021-04-19 Due date for @marionlouveaux: 2021-04-27
Version accepted: TBD

---

```
Package: tidytags
Version: 0.1.0
Title: Simple Collection and Powerful Analysis of Twitter Data
    person("K. Bret", "Staudt Willet", ,
      email = "bret@bretsw.com", role = c("aut", "cre"),
      comment = c(ORCID = "0000-0002-6984-416X")
    ),
    person("Joshua M.", "Rosenberg", ,
      role = c("aut"),
      comment = c(ORCID = "0000-0003-2170-0447")
    )
  )
Description: {tidytags} coordinates the simplicity of collecting tweets over time
    with a [Twitter Archiving Google Sheet](https://tags.hawksey.info/) (TAGS) and the utility of the
    [{rtweet} package](https://rtweet.info/) for processing and preparing additional Twitter metadata.
    {tidytags} also introduces functions developed to facilitate systematic yet
    flexible analyses of data from Twitter.
    dplyr (>= 0.8),
    googlesheets4 (>= 0.2),
    purrr (>= 0.3),
    readr (>= 1.3),
    rlang(>= 0.4),
    rtweet (>= 0.7),
    stringr (>= 1.4),
    tibble (>= 3.0),
    tidyr (>= 1.0),
    tidyselect (>= 1.0)
    beepr,
    covr,
    longurl,
    mapsapi,
    mapview,
    tidyverse,
    urltools,
    usethis
```

	- [ ] workflow automataion

{tidytags} allows for both simple data collection and thorough data analysis. In short, {tidytags} first uses a [Twitter Archiving Google Sheet](https://tags.hawksey.info/) (TAGS) to easily collect tweet ID numbers and then uses the R package {rtweet} to re-query the Twitter API to collect additional metadata. {tidytags} also introduces new functions developed to facilitate systematic yet flexible analyses of data from Twitter.

The target users for {tidytags} are social scientists (e.g., educational researchers) who have an interest in studying Twitter data but are relatively new to R, data science, or social network analysis. {tidytags} scaffolds tweet collection and analysis through a simple workflow that still allows for robust analyses.

{tidytags} wraps together functionality from several useful R packages, including {googlesheets4} to bring data from the TAGS tracker into R and {rtweet} for retrieving additional tweet metadata. The contribution of {tidytags} is to bring together the affordance of TAGS to easily collect tweets over time (which is not straightforward with {rtweet}) and the utility of {rtweet} for collecting additional data (which are missing from TAGS). Finally, {tidytags} reshapes data in preparation for geolocation and social network analyses that should be accessible to relatively new R users.
