### 2020-10-05 #382 @maelle
:wave: @bretsw, some first remarks from me (another editor).

* I am generally wondering how the reviewers can test the package. For instance is the TAGS used in the vignette usable by anyone? I see the vignette chunks aren't evaluated, why is it so? Not that I have anything against pre-compiling vignettes in such a case!

* Could you add documentation on data protection? See https://devguide.ropensci.org/policies.html#ethics-data-privacy-and-human-subjects-research I saw a related open issue in your repo.

* In the pkgdown reference page, it would make sense to [group functions](https://pkgdown.r-lib.org/reference/build_reference.html#reference-index).

* The README content could be partly re-used as a Get Started vignette (just name a vignette tidytags.Rmd). [How to re-use chunks](https://www.garrickadenbuie.com/blog/dry-vignette-and-readme/).

* Relatedly, although the README structure is clear, I'd appreciate a setup checklist to provide an overview. Or  maybe a setup vignette like the rtweet vignette about secrets.

```markdown

* Have a TAGS with blablabla ready (blablabla being the public URL)
* Setup rtweet credentials if you want to use blablabla
* Setup Google geocoding credentials if you want to use blablabla

```

* In the TAGS setup explanation maybe some screenshots would make sense. I say maybe as a) they easily go stale b) your text seems quite clear (not tested yet) and a screenshot wouldn't _replace_ instructions.

* Regarding tests am I correct that they're skipped everywhere but locally? Is it both because you are using authentication & Twitter data you don't own/can't share? I am actually working on the ["HTTP testing in R" book](https://github.com/ropensci-books/http-testing) so I'd recommend choosing an approach with some sort of fake data. I can help more once I know more about your constraints.
