### 2020-10-24 #382 @bretsw
@maelle, thank you again for your feedback! Below are responses from @jrosen48 and I, and we have pushed all changes to the repo: https://github.com/bretsw/tidytags

We're looking forward to more dialogue on this!
//Bret

> 1. I am generally wondering how the reviewers can test the package. For instance is the TAGS used in the vignette usable by anyone? I see the vignette chunks aren't evaluated, why is it so? Not that I have anything against pre-compiling vignettes in such a case!

**Our response:** The example vignette is usable by anyone. We’ve just set the chunks as `eval=FALSE` because many of the processes are slow, even when limiting the examples to small data. But all the code should work and anyone can view/pull data from the TAGS tracker linked  to in the vignette.

> 2. Could you add documentation on data protection? See https://devguide.ropensci.org/policies.html#ethics-data-privacy-and-human-subjects-research I saw a related open issue in your repo.

**Our response:** We added several paragraphs to a new “Considerations Related to Ethics, Data Privacy, and Human Subjects Research” in the README.
Regarding the open issue in the repo (Issue #2), we are  considering this enhancement for a future version of {tidytags}.

> 3. In the pkgdown reference page, it would make sense to group functions.

**Our response:** We have grouped the functions. See https://bretsw.github.io/tidytags/reference/index.html

> 4. The README content could be partly re-used as a Get Started vignette (just name a vignette tidytags.Rmd). How to re-use chunks.

**Our response:** We have created a “Getting started with tidytags” vignette (https://bretsw.github.io/tidytags/articles/setup.html).

> 5. Relatedly, although the README structure is clear, I'd appreciate a setup checklist to provide an overview. Or maybe a setup vignette like the rtweet vignette about secrets.

**Our response:** We have added duplicate checklists in both the  README and in the new “Getting started with tidytags” vignette.

> 6. In the TAGS setup explanation maybe some screenshots would make sense. I say maybe as a) they easily go stale b) your text seems quite clear (not tested yet) and a screenshot wouldn't replace instructions.

**Our response:** We have added screenshots to the “Getting started with tidytags” vignette.

> 7. Regarding tests am I correct that they're skipped everywhere but locally? Is it both because you are using authentication & Twitter data you don't own/can't share? I am actually working on the "HTTP testing in R" book so I'd recommend choosing an approach with some sort of fake data. I can help more once I know more about your constraints.

**Our response:** We have set up tests for many of the functions, often using fake data. The functions that we cannot test for CRAN/Travis (i.e., we can only test locally) are those that are querying either the Twitter or  Google Maps API. For example, `add_users_data()` uses a Twitter API key, so we cannot test all aspects of it for CRAN or Travis.
