### 2020-10-24 #382 @maelle
Is there no free tier for Google geocoding? If so then my opencage suggestion is becoming more important. :thinking:
