### 2020-10-24 #382 @maelle
well i guess the free credits one gets when creating the billing account count as free tier maybe :thinking:
