### 2020-11-25 #382 @maelle
Thanks a ton for all your work!

* Regarding the code of conduct, awesome! Note that after acceptance the rOpenSci COC will apply cf https://devguide.ropensci.org/collaboration.html#coc-file

* I am glad about OpenCage, not only because it's a nice package :grin: but also because setup is so much easier.

* Speaking of setup docs, they are really good!

* Now regarding the contributing guide what's missing is a sandbox TAGS like OSF's development environment/ODK sandbox. Could reviewers (and the editor?) get access to a TAGS used for testing the package? (even if ideally once I start looking for reviewers I hope to find TAGS users who'd use the package on their own data)

* Speaking of testing. You will want both tests that use cached responses of some sort, and tests making real requests at least once in a while.
    * The rOpenSci "HTTP testing in R" book is undergoing a major update (by me :slightly_smiling_face: ) in particular I've worked on demos of the three tools for HTTP testing there are in R. See https://github.com/ropensci-books/http-testing/pull/47 and the Netlify preview linked from the PR checks https://5fbb93389e79d3867339005d--http-testing-book.netlify.app/ (the Whole games section is the one with the demos)
    * Regarding tests with real requests especially on CI the best guidance out there is https://gargle.r-lib.org/articles/articles/managing-tokens-securely.html
    * If I get access to a sandbox TAGS I can help with setting up tests, e.g. making a PR adding a first vcr test if that helps.
