### 2020-12-11 #382 @maelle
reg testthat's new version https://testthat.r-lib.org/articles/third-edition.html
