### 2020-12-16 #382 @bretsw
Hi @maelle! Thanks again for pointing me in all the right directions. I _think_ I've gotten vcr tests working with Twitter and OpenCage APIs. At least it looks good on my local machine. I just pushed a big update so we'll see. When I ran `covr::package_coverage()` locally it looked like the tests are covering around 85% now. I've gotten tests for all functions, in any case.

It does look like R-CMD-check is failing on Github though.

One place where I got stuck and couldn't find my way through was on a vcr test for `get_url_domain()`. The vcr test worked with long URLS but not shortened ones (e.g., bit.ly). Seems like there's something going on in a dependency package that I'm not catching yet. Anyway, set those test to `skip()` for now.

Let me know what's next! THANK YOU, as always, for all your help.
