### 2020-12-17 bretsw/tidytags#30 @maelle
Twitter token in vcr cassettes
When you use Twitter API the secret is not the path to the Twitter httr token (what's in the "secret" `TWITTER_PAT` environment variable) but inside the token file.

In such a case you need to use https://books.ropensci.org/http-testing/vcr-security.html#if-the-secret-is-in-a-request-header

* Please change your vcr configuration
* Please invalidate your token that has been leaked. Probably via de-authorizing the Twitter app you created for getting the token.
