### 2020-12-17 #382 @maelle
I see a security problem, I've opened an issue, please look at this urgently https://github.com/bretsw/tidytags/issues/30 @bretsw (at least inactivate the token that was leaked).

Feel free to tell me where in the docs we could have warned you more (this coming from someone who leaked their own GitHub PAT a few weeks ago :grimacing: )
