### 2020-12-17 bretsw/tidytags#30 @bretsw
Thank you for catching this. I've revoked the leaked tokens and generated new ones.
