### 2020-12-17 #382 @maelle
A few last comments for today

* the current vignette building approach is fine, but for info with `vcr::inject_cassette()`/`vcr::eject_cassette()` (and the proper vcr security configuration) you could instead use vcr cassettes in the vignettes.
* you might want to take some notes around your test setup in CONTRIBUTING.md, see https://books.ropensci.org/http-testing/contributor-friendliness.html
* in the contributing guide what does "common sense" mean?
* the checks in my typo fix PR failed, could you ensure tests pass for forks? I suspect it's because there's no fake Twitter httr token file around. Maybe something like below can help

```r
httr::oauth2.0_token(
  endpoint = httr::oauth_endpoints("twitter"),
  app = httr::oauth_app("foobar", "foobar", "foobar"),
  credentials = list(access_token = "foobar")
)
```

I might hold off looking for reviewers until January (I'll be on vacation for two weeks, and I suspect it'll be hard to get responses). In any case you've done awesome work on the package, thank you.
