### 2020-12-17 #382 @maelle
(I see the checks don't pass for the repo either, hence the coverage badge still showing low coverage?)
