### 2021-01-28 #382 @bretsw
> > One place where I got stuck and couldn't find my way through was on a vcr test for get_url_domain(). The vcr test worked with long URLS but not shortened ones (e.g., bit.ly). Seems like there's something going on in a dependency package that I'm not catching yet. Anyway, set those test to skip() for now.
>
> This might be worth a question in vcr repo / in rOpenSci forum.

Opened an issue in the vcr repo: https://github.com/ropensci/vcr/issues/220
