### 2021-03-02 #382 @bretsw
I've updated our codecov too and tidytags is at 88% coverage.

Also, I've made a number of updates trying to get CI tests to work. I updated `geocode_tags()` to reflect the new function from OpenCage, `oc_forward_df()` which does exactly what I had baked into my older code.

So, CI tests are currently passing on Mac and Ubuntu but the Windows test is returning a strange error:
`Error:   Package suggested but not available: 'ggraph'`

Have you seen something like this before? The ggraph package definitely still exists and is usable on my local machine (and apparently on Mac and Ubuntu generally). Any ideas there?
