### 2021-03-02 #382 @maelle
Thanks for the update, this is awesome! (and I'd say a tribute to @dpprdan's new opencage functions)

Usually when I see such errors it happens right after a package update on CRAN, and so the binary is not available. It doesn't seem to be the case here.

I was thinking about suggesting you tweak the workflow so that remotes would error at the dependency installation step, not later (no need to do the check if not all dependencies are present) but then I remembered the tidyverse team seems to be switching their workflows to using pak instead: https://github.com/r-lib/pkgdown/blob/9c95f00d2505ddd83c4722e019202370715d3a3d/.github/workflows/R-CMD-check.yaml#L54

I think using pak instead of remotes might be a good idea and with a bit of luck you'd get no error / a more informative failure on Windows?
