### 2021-03-03 #382 @bretsw
Hi @maelle, I'm working on setting up real tests for CI, and I wanted to run my plan by you to see what you think.

I've created a new file, "weekly-check.yaml" for the real tests. I've scheduled the tests to occur once every week at midnight Sundays with `- cron:  '0 0 * * Sun'`.

tidytags functions want to find API keys from environmental variables, so I've stored opencage and rtweet keys in GitHub Secrets. I'll add them then with things like `OPENCAGE_KEY: ${{ secrets.OPENCAGE_KEY }}` and `TWITTER_API_KEY: ${{ secrets.TWITTER_API_KEY }}` (there are five secrets for Twitter, as listed at https://docs.ropensci.org/rtweet/articles/auth.html). rtweet will create a Twitter token with `rtweet::create_token()` so I'm adding that to the steps in the .yaml file. The token gets stored in (and then automatically accessed through) an .rds file. I've added "*.rds" to .gitignore so the token in the .rds file shouldn't be exposed at all.

I think that's it. What do you think? Is this a good plan? Am I missing something obvious that will unwittingly expose a secret?
