### 2021-03-05 #382 @maelle
You mean https://docs.ropensci.org/rtweet/articles/auth.html#2-access-token-secret-method? That sounds great and I should document this in the book as it sounds easier than encrypting.

As this thread taught us both how many ways there are to leak secrets 😅 (not your fault!): the token will be in an app dir or so, **not** in the current directory, correct? Just to make sure it won't get leaked in a check .tar.gz in case of failured.

Also, can you confirm you use a special Twitter account for the app and token, not your account?
