### 2021-03-08 #382 @bretsw
Hi, @maelle. I set up a special Twitter account for the app and token, and the real CI tests are queued to run in about 5 minutes. So, we'll see.

In the meantime, now the regular build has failed all CI tests: https://github.com/bretsw/tidytags/runs/2056401038

Is this because R updated to v4.1 on Friday 3/5? Any advice on how I can help tidytags catch up?
