### 2021-03-08 #382 @bretsw
> In some of the runs it seems the error comes from vcr. Could your tidytags example sheet have changed somehow?
>
> Btw the link to that sheet contains a key https://github.com/bretsw/tidytags/runs/2056400999#step:10:142 Is this expected?

Nope, this was not expected. That was my Google API key, which is now deleted. I need to dig into this, because the Google API key was only used back when tidytags used Google Maps for geocoding. I would not have expected that key to be called at all since we switched to OpenCage.
