### 2021-03-17 #382 @bretsw
I though I had this Google API key issue figured out, but no such luck. @maelle, how familiar are you with googlesheets4?

Something is happening with the stored Google API key that I'm not understanding. Somehow my old Google API key is stored somewhere in a way that I can't seem to change or access, until I record vcr cassettes and see that it is exposed in the request URL. Which should be fine, because I've revoked that token and now have a new one. However, my new API key won't work when I run `googlesheets4::gs4_auth_configure(api_key = Sys.getenv("GOOGLE_API_KEY"))`, but running `googlesheets4::gs4_deauth()` (which sets the API token to NULL) and `googlesheets4::gs4_auth_configure(api_key = NULL)` (which sets the API key to NULL) somehow lets  me query the sheets API. That is, with both a NULL key and a NULL token, I can successfully run `googlesheets4::range_read(googlesheets4::gs4_examples("deaths"))` or perform my tidytags package tests (locally).

In sum, there's an old, deactivated API key stored somewhere I can't locate and being accessed in a way I can't decipher. The old API key is still currently exposed in several vcr cassettes (in the "fixtures" directory), but I'm ok with this for now because the key is actually decommissioned.

Any ideas?
