### 2021-03-18 #382 @maelle
Not familiar at all!

If I follow correctly there are two problems

* reg the API key in vcr cassettes you need to tweak vcr configuration. https://docs.ropensci.org/vcr/articles/configuration.html#filter_query_parameters Like for request headers, with vcr latest version (not on CRAN yet though) you can filter query parameters.
* now as to where that key is saved, could you try searching for the key string on your computer? Could https://googlesheets4.tidyverse.org/reference/gs4_auth_configure.html help?

The next step would be to ask for help on RStudio community forum (since googlesheets4 is an RStudio package, I'd expect more users there than on rOpenSci forum).
