### 2021-03-22 #382 @maelle
Hello! It seems the problem is not the cron syntax but your referring to `matrix.config.os` without defining it (it can't be shared betwen workflow files).
