### 2021-03-22 #382 @bretsw
Isn't the matrix defined on [lines 28-30](https://github.com/bretsw/tidytags/blob/81a5eeac2bf9a60ccb85f2beee547c485bde28c6/.github/workflows/weekly-check.yaml#L28)? This reflects[ lines 26-28](https://github.com/rladies/meetupr/blob/c9c584335377966c7c9e08534089f84d921cff50/.github/workflows/with-auth.yaml#L26) in the rladies' meetupr yaml.
