### 2021-03-22 #382 @maelle
Right but it seems it isn't parsed? The meetupr YAML doesn't work either :sweat_smile: https://github.com/rladies/meetupr/actions/runs/673097825
