### 2021-03-22 #382 @bretsw
Thanks @dpprdan for catching that this actually _did_ run (yay!) and @maelle for demonstrating the solution for parsing in meetupr. I'm testing with tidytags now and will report back. I really appreciate the community support!
