### 2021-03-22 #382 @bretsw
Hi @maelle, looks like everything is working with tidytags:

- CI tests work [with vcr](https://github.com/bretsw/tidytags/actions/runs/676301323) when I push changes.
- CI tests work [with real requests](https://github.com/bretsw/tidytags/actions/runs/676464932), now scheduled for every Monday, Wednesday, and Friday

I think(?) I've checked everything off the list!
