### 2021-03-29 #382 @maelle
@bretsw please don't forget to add the badge mentioned in https://github.com/ropensci/software-review/issues/382#issuecomment-804893607 :slightly_smiling_face:
