### 2021-03-29 #382 @bretsw
Thanks for the reminder, I totally missed that prompt, probably from skimming past message from the review bot. Sorry bot! I'll add NEWS.md next.
