### 2021-04-04 #382 @llrs
The setup vignette is not clear what steps are necessary and which not, I would suggests adding titles and an index.

 he vignettes have a "Pain point #4" which I couldn't find referenced anywhere. Also, perhaps using the more descriptive title would make it easier for users to know what it is about (Removing the Pain point reference entirely of the title of the section). Not that they aren't pain points but just redirect users when needed to the solutions/documentation as they go.
However, most of the code chunks of the vignette are not run (as reported by BiocCheck):

```
    * WARNING: Evaluate more vignette chunks.
        # of code chunks: 8
        # of eval=FALSE: 5
        # of nonexecutable code chunks by syntax: 0
        # total unevaluated 5 (62%)
```

And of those run are adding documentation or set up of the vignette. Perhaps some kind of setup specific for the vignettes could be used, otherwise they defeat their purpose and turn into plain READMEs. (I know it is not easy for CRAN, so maybe set them up as articles just on the website but outside CRAN?)

To create the google API key step is not clear enough (perhaps a redesign on API configuration  interface?). An indication to use Google Sheet API to the question "Find out what kind of credentials you need?", would be helpful.

It should be pointed out that OpenCage Geocoding API key is not needed to use the package. Also the discussion about the price and API limits might be good for an issue but doesn't fit well on the vignette (I've seen that @maelle asked for this, but now that the package is settle in Open Cage maybe it is no longer needed or it can be reduced).

On the chunk about "dplyr::glimpse(example_after_rtweet)" I get a different result 2204 rows compared to the 2,215 reported on the vignette.

When I run the  following code chunk I get an error (as I don't have the package longurl yet)

```r
example_domains <- get_url_domain(example_urls)
```

Before using a package in Suggests, it should be tested if they can be loaded (you can use `rlang::is_installed(longurl` or `requireNamespace("longurl", quietly = TRUE)`).

Last, I don't know how to push data to the google sheet TAGS created on the first vignette.

Name of the function repeated on the description on `add_users_data`, I think it is not needed.
- [x] **Examples** (that run successfully locally) for all exported functions

Examples cannot be run without the authentication setup and there is no mention of this on the help pages. Perhaps a minor comment will remind users.

There isn't any BugReports but on the vignettes there is info about how to get help.
Would suggest to add the issues link on the description too. The contributing file is extensive and well organized.

##### For packages co-submitting to JOSS

- [ ] The package has an **obvious research application** according to [JOSS's definition](https://joss.theoj.org/about#submission_requirements)

The authors say: "Yet, many approaches to collecting social media data in the moment require important technical skill that may dissuade social scientists from getting started". However this packages requires authentication in 2 or 3 API and a special google sheets. I have some doubts tidytags will attract the target audience.
 None of the references have dois or urls to online resources.

There's an additional `"` on the yaml heading of paper.md that prevented viewing the paper.

I got 2 tests that failed and 3 warnings (besides 2 that were skipped). The test-get_url_domain.R:14:5 test reported `domain4` not equal to "npr.org", on the browser I get asked for cookie consent on the browser, when run locally, outside testthat or vcr, I get the url of choice.npr.org.
The other failing test are weird (as I don't get them when I run them on the R console but only on the build/check Rstudio panel).

I have a development version installed of vcr and one of the warnings is related to it. The new version warns when the cassettes are empty, this in my experience means that the test it not conclusive, but this could also be related to not having the  geo code API enabled.
The other warnings are on test-get_url_domain.R, lines 3 and 32, Invalid URL I'm not sure why, because when I paste on my browser I get redirected to https://www.aect.org/about_us.php. (BTW perhaps the link can be changed to https instead of http).

---

The package rtweet is experimenting drastic changes (I'm involved on rtweet maintenance) and there will be a major release with breaking changes. Probably it will break this package (the recommendation about the token will change for instance), so be ready to update it accordingly.

The package contains relative few simple functions that provide more data from twitter or make it easier to analyze it. I have not analyzed Twitter data (except for an overview of a user account), so don't know how useful the data is. I am not a user of TAGs but I'm a bit puzzles how to add information to the google sheet: if I'm a new user how should I do it? I mean I can get the template but how I fill it? I think this package would be easier for non technical people if it included a function to add the information gathered via rtweet or processed with the package back to the original google sheet.

Haven't fully read the paper.md for JOSS but I think it is short enough and comprenhsive of the package.

From a more technical point of view, I have some comments about the code and the package:

There are 75 lines longer than 80 characters, try to reduce them. Probably it is just a matter of style and perhaps creating new shorter variables

Also namespaces in Imports field not imported from: ?gargle? ?readr?. All declared Imports should be used.

The `get_char_tweet_ids` function could be improved, with only one argument if it is a data.frame then extract the status_id and get the ID via id_str. If it is an url you can just extract the last numbers with `gsub("https?\\://twitter.com\\/.+/statuses/", "", df$status_url)`, no need to modify the data.frame and then extract the vector again.

On `process_tweets` you can simplify the is_self_reply to `ifelse(.data$is_reply & .data$user_id == .data$reply_to_user_id, TRUE, FALSE)`.

On `get_upstream_replies` the examples are not informative, as there are no replies to get data from on the example dataset. You make multiple calls to `pull_tweet_data`, some of them might be unnecessary. The `process_tweets` can be called just once at the end instead of multiple times and on each loop run. This should speed up the process. Also, if there are at most 90000 tweets taken from each run, then you can estimate the number of iterations needed and inform the user. This might make the wait easier. Perhaps it would be better to use `lookup_many tweets` as it does a similar process. However, users might hit the rate limit and I don't see any information being passed to the user regarding this.

Looking at `create_edgelist`, it calls `process_tweets` and also `get_replies`, `get_retweets`, `get_quotes`, `get_mentions` which they call `process_tweets` too. Perhaps some internal functions could be created to avoid calling `process_tweets` multiple times on the same data.
