### 2021-04-05 #382 @maelle
Thanks a lot for your review @llrs! :rocket:

Note that regarding JOSS, we've just changed the process as JOSS will be the ones determining whether the software fits in their scope.
