### 2021-04-22 #382 @maelle
@llrs which rtweet version did you use for your review, by the way?

@bretsw do you use rtweet CRAN version or the GitHub version with the newer changes?

Thank you :slightly_smiling_face:
