### 2021-04-23 #382 @marionlouveaux
<details> <summary><strong>Session Info</strong></summary>

```
    > sessionInfo()
R version 4.0.4 (2021-02-15)
Platform: x86_64-pc-linux-gnu (64-bit)
Running under: Ubuntu 20.04.2 LTS

Matrix products: default
BLAS:   /usr/lib/x86_64-linux-gnu/blas/libblas.so.3.9.0
LAPACK: /usr/lib/x86_64-linux-gnu/lapack/liblapack.so.3.9.0

locale:
 [1] LC_CTYPE=fr_FR.UTF-8       LC_NUMERIC=C               LC_TIME=fr_FR.UTF-8        LC_COLLATE=fr_FR.UTF-8     LC_MONETARY=fr_FR.UTF-8
 [6] LC_MESSAGES=fr_FR.UTF-8    LC_PAPER=fr_FR.UTF-8       LC_NAME=C                  LC_ADDRESS=C               LC_TELEPHONE=C
[11] LC_MEASUREMENT=fr_FR.UTF-8 LC_IDENTIFICATION=C

attached base packages:

other attached packages:
[1] tidytags_0.1.2 devtools_2.3.2 usethis_2.0.1  magrittr_2.0.1

  [1] uuid_0.1-4              systemfonts_1.0.1       igraph_1.2.6            lazyeval_0.2.2          sp_1.4-5                crosstalk_1.1.1
  [7] leaflet_2.0.4.1         ggplot2_3.3.3           urltools_1.7.3          digest_0.6.27           leafpop_0.0.6           htmltools_0.5.1.1
 [13] viridis_0.5.1           leaflet.providers_1.9.0 fansi_0.4.2             memoise_2.0.0           covr_3.5.1              googlesheets4_0.3.0
 [19] remotes_2.2.0           readr_1.4.0             graphlayouts_0.7.1      svglite_2.0.0           askpass_1.1             prettyunits_1.1.1
 [25] colorspace_2.0-0        ggrepel_0.9.1           rtweet_0.7.0            xfun_0.22               dplyr_1.0.5             leafem_0.1.3
 [31] callr_3.6.0             crayon_1.4.1            jsonlite_1.7.2          roxygen2_7.1.1          attachment_0.2.1        brew_1.0-6
 [37] glue_1.4.2              xmlparsedata_1.0.5      polyclip_1.10-0         gtable_0.3.0            gargle_1.1.0            webshot_0.5.2
 [43] pkgbuild_1.2.0          scales_1.1.1            DBI_1.1.1               opencage_0.2.2          Rcpp_1.0.6              viridisLite_0.3.0
 [49] units_0.7-1             proxy_0.4-25            praise_1.0.0            clisymbols_1.2.0        stats4_4.0.4            xopen_1.0.0
 [55] htmlwidgets_1.5.3       rex_1.2.0               httr_1.4.2              RColorBrewer_1.1-2      ellipsis_0.3.1          pkgconfig_2.0.3
 [61] farver_2.1.0            sass_0.3.1              utf8_1.2.1              crul_1.1.0              labeling_0.4.2          tidyselect_1.1.0
 [67] rlang_0.4.10            munsell_0.5.0           cellranger_1.1.0        tools_4.0.4             cachem_1.0.4            cli_2.4.0
 [73] generics_0.1.0          evaluate_0.14           stringr_1.4.0           fastmap_1.1.0           yaml_2.2.1              processx_3.5.0
 [79] knitr_1.31              fs_1.5.0                tidygraph_1.2.0         purrr_0.3.4             satellite_1.0.2         ggraph_2.0.5
 [85] xml2_1.3.2              compiler_4.0.4          rstudioapi_0.13         curl_4.3                png_0.1-7               e1071_1.7-6
 [91] testthat_3.0.2          tibble_3.1.0            tweenr_1.0.2            bslib_0.2.4             stringi_1.5.3           pkgreviewr_0.2.0
 [97] cyclocomp_1.1.0         ps_1.6.0                desc_1.3.0              lattice_0.20-41         whoami_1.3.0            classInt_0.4-3
[103] goodpractice_1.0.2      vctrs_0.3.7             pillar_1.6.0            lifecycle_1.0.0         triebeard_0.3.0         jquerylib_0.1.3
[109] rcmdcheck_1.3.3         raster_3.4-5            mapview_2.9.0           R6_2.5.0                KernSmooth_2.23-18      gridExtra_2.3
[115] sessioninfo_1.1.1       codetools_0.2-18        MASS_7.3-53.1           assertthat_0.2.1        pkgload_1.2.0           openssl_1.4.3
[121] rprojroot_2.0.2         withr_2.4.1             httpcode_0.3.0          hms_1.0.0               lintr_2.0.1             grid_4.0.4
[127] tidyr_1.1.3             class_7.3-18            rmarkdown_2.7           googledrive_1.0.1       sf_0.9-8                ggforce_0.3.3
[133] base64enc_0.1-3         ratelimitr_0.4.1
```

## Test installation

Local installation took several minutes (approx. 3 to 5 minutes) because there is many dependencies. On my machine, it had to install 35 packages
+ 3 dependencies.

<details> <summary><strong>Installation details</strong></summary>

```
Installing 35 packages: fauxpas, selectr, broom, forcats, uuid, ids, googledrive, gargle, data.table, blob, polyclip, tweenr, webmockr, rvest, modelr, haven, googlesheets4, dtplyr, dbplyr, ratelimitr, satellite, leafpop, graphlayouts, tidygraph, ggrepel, ggforce, audio, vcr, tidyverse, opencage, mapview, longurl, ggraph, beepr, rtweet
Installing packages into ‘/home/marion/R/x86_64-pc-linux-gnu-library/4.0’
(as ‘lib’ is unspecified)
also installing the dependencies ‘cli’, ‘lubridate’, ‘pillar’
```

## Check package integrity

### run checks on `tidytags` source

Recommendation: In Contributing.md, remind potential contributors to follow the [Getting started with tidytags guide](https://bretsw.github.io/tidytags/articles/setup.html) before proceeding to a check on the package. Without the API keys, it doesn’t
work.

**7 failed tests**, all related to vcr. These tests pass if I delete the fixtures folder. NB: the errors message contain my secret tokens for
Twitter, so I removed most of the URLs and replaced it by “………..”.

<details> <summary><strong>Error message for failed tests</strong></summary>

    ── Error (test-add_users_data.R:14:3): user data is added properly ─────────────
    Error:

    ================================================================================
    An HTTP request has been made that vcr does not know how to handle:
    GET https://api.twitter.com/1.1/users/lookup.json?screen_name=gsa_aect%2CAECT...........
    vcr is currently using the following cassette:
      - ../fixtures/users_info.yml
        - record_mode: once
        - match_requests_on: method, uri
    Set `VCR_VERBOSE_ERRORS=TRUE` for more verbose errors
    If you're not sure what to do, open an issue https://github.com/ropensci/vcr/issues
    & see https://books.ropensci.org/http-testing
    ================================================================================

    Backtrace:
         █
      1. ├─vcr::use_cassette(...) test-add_users_data.R:14:2
      2. │ └─cassette$call_block(...)
      3. └─tidytags::add_users_data(el) test-add_users_data.R:16:4
      4.   └─rtweet::lookup_users(all_users)
      5.     ├─base::do.call("lookup_users_", args)
      6.     └─rtweet:::lookup_users_(...)
      7.       └─rtweet:::.user_lookup(users, token)
      8.         └─rtweet:::TWIT(get = get, url, token)
      9.           └─httr::GET(url, ...)
     10.             └─httr:::request_perform(req, hu$handle$handle)
     11.               └─httr:::perform_callback("request", req = req)
     12.                 └─webmockr:::callback(...)
     13.                   └─webmockr::HttrAdapter$new()$handle_request(req)
     14.                     └─private$request_handler(req)$handle()
     15.                       └─eval(parse(text = req_type_fun))(self$request)
     16.                         └─err$run()
     17.                           └─self$construct_message()
    ── Error (test-get_upstream_replies.R:2:3): get_upstream_replies() finds additional replies ──
    Error:

    ================================================================================
    An HTTP request has been made that vcr does not know how to handle:
    POST https://api.twitter.com/1.1/statuses/lookup.json?id=...............
    vcr is currently using the following cassette:
      - ../fixtures/upstream_replies.yml
        - record_mode: once
        - match_requests_on: method, uri
    Set `VCR_VERBOSE_ERRORS=TRUE` for more verbose errors
    If you're not sure what to do, open an issue https://github.com/ropensci/vcr/issues
    & see https://books.ropensci.org/http-testing
    ================================================================================

    Backtrace:
         █
      1. ├─vcr::use_cassette(...) test-get_upstream_replies.R:2:2
      2. │ └─cassette$call_block(...)
      3. └─tidytags::pull_tweet_data(tags_data) test-get_upstream_replies.R:4:4
      4.   ├─base::ifelse(...)
      5.   ├─base::ifelse(...)
      6.   └─rtweet::lookup_statuses(get_char_tweet_ids(df[1:n, ]))
      7.     ├─base::do.call("lookup_statuses_", args)
      8.     └─rtweet:::lookup_statuses_(...)
      9.       └─rtweet:::.status_lookup(statuses[from:to], token = token)
     10.         └─rtweet:::TWIT(get = get, url, token)
     11.           └─httr::POST(url, ...)
     12.             └─httr:::request_perform(req, hu$handle$handle)
     13.               └─httr:::perform_callback("request", req = req)
     14.                 └─webmockr:::callback(...)
     15.                   └─webmockr::HttrAdapter$new()$handle_request(req)
     16.                     └─private$request_handler(req)$handle()
     17.                       └─eval(parse(text = req_type_fun))(self$request)
     18.                         └─err$run()
     19.                           └─self$construct_message()
    ── Error (test-get_upstream_replies.R:34:3): get_upstream_replies() works with no new replies found ──
    Error:

    ================================================================================
    An HTTP request has been made that vcr does not know how to handle:
    GET https://api.twitter.com/1.1/statuses/lookup.json?id=NA&tweet_mode=extended..............
    vcr is currently using the following cassette:
      - ../fixtures/upstream_replies_empty.yml
        - record_mode: once
        - match_requests_on: method, uri
    Set `VCR_VERBOSE_ERRORS=TRUE` for more verbose errors
    If you're not sure what to do, open an issue https://github.com/ropensci/vcr/issues
    & see https://books.ropensci.org/http-testing
    ================================================================================

    Backtrace:
         █
      1. ├─vcr::use_cassette(...) test-get_upstream_replies.R:34:2
      2. │ └─cassette$call_block(...)
      3. └─tidytags::get_upstream_replies(sample_data) test-get_upstream_replies.R:35:4
      4.   ├─base::nrow(pull_tweet_data(id_vector = unknown_replies$reply_to_status_id))
      5.   └─tidytags::pull_tweet_data(id_vector = unknown_replies$reply_to_status_id)
      6.     ├─base::ifelse(...)
      7.     ├─base::ifelse(...)
      8.     └─rtweet::lookup_statuses(id_vector[1:n])
      9.       ├─base::do.call("lookup_statuses_", args)
     10.       └─rtweet:::lookup_statuses_(...)
     11.         └─rtweet:::.status_lookup(statuses[from:to], token = token)
     12.           └─rtweet:::TWIT(get = get, url, token)
     13.             └─httr::GET(url, ...)
     14.               └─httr:::request_perform(req, hu$handle$handle)
     15.                 └─httr:::perform_callback("request", req = req)
     16.                   └─webmockr:::callback(...)
     17.                     └─webmockr::HttrAdapter$new()$handle_request(req)
     18.                       └─private$request_handler(req)$handle()
     19.                         └─eval(parse(text = req_type_fun))(self$request)
     20.                           └─err$run()
     21.                             └─self$construct_message()
    ── Error (test-lookup_many_tweets.R:3:3): lookup_many_tweets() retrieves additional metadata like pull_tweet_data() ──
    Error:

    ================================================================================
    An HTTP request has been made that vcr does not know how to handle:
    GET https://api.twitter.com/1.1/statuses/lookup.json?id=................
    vcr is currently using the following cassette:
      - ../fixtures/lookup_many.yml
        - record_mode: once
        - match_requests_on: method, uri
    Set `VCR_VERBOSE_ERRORS=TRUE` for more verbose errors
    If you're not sure what to do, open an issue https://github.com/ropensci/vcr/issues
    & see https://books.ropensci.org/http-testing
    ================================================================================

    Backtrace:
         █
      1. ├─vcr::use_cassette(...) test-lookup_many_tweets.R:3:2
      2. │ └─cassette$call_block(...)
      3. └─tidytags::pull_tweet_data(sample_tags, n = 10) test-lookup_many_tweets.R:5:4
      4.   ├─base::ifelse(...)
      5.   ├─base::ifelse(...)
      6.   └─rtweet::lookup_statuses(get_char_tweet_ids(df[1:n, ]))
      7.     ├─base::do.call("lookup_statuses_", args)
      8.     └─rtweet:::lookup_statuses_(...)
      9.       └─rtweet:::.status_lookup(statuses[from:to], token = token)
     10.         └─rtweet:::TWIT(get = get, url, token)
     11.           └─httr::GET(url, ...)
     12.             └─httr:::request_perform(req, hu$handle$handle)
     13.               └─httr:::perform_callback("request", req = req)
     14.                 └─webmockr:::callback(...)
     15.                   └─webmockr::HttrAdapter$new()$handle_request(req)
     16.                     └─private$request_handler(req)$handle()
     17.                       └─eval(parse(text = req_type_fun))(self$request)
     18.                         └─err$run()
     19.                           └─self$construct_message()
    ── Error (test-pull_tweet_data.R:7:3): pull_tweet_data() is able to retrieve additional metadata starting with dataframe ──
    Error:

    ================================================================================
    An HTTP request has been made that vcr does not know how to handle:
    GET https://api.twitter.com/1.1/statuses/lookup.json?id=.................
    vcr is currently using the following cassette:
      - ../fixtures/metadata_from_df.yml
        - record_mode: once
        - match_requests_on: method, uri
    Set `VCR_VERBOSE_ERRORS=TRUE` for more verbose errors
    If you're not sure what to do, open an issue https://github.com/ropensci/vcr/issues
    & see https://books.ropensci.org/http-testing
    ================================================================================

    Backtrace:
         █
      1. ├─vcr::use_cassette(...) test-pull_tweet_data.R:7:2
      2. │ └─cassette$call_block(...)
      3. └─tidytags::pull_tweet_data(sample_tags, n = 10) test-pull_tweet_data.R:8:4
      4.   ├─base::ifelse(...)
      5.   ├─base::ifelse(...)
      6.   └─rtweet::lookup_statuses(get_char_tweet_ids(df[1:n, ]))
      7.     ├─base::do.call("lookup_statuses_", args)
      8.     └─rtweet:::lookup_statuses_(...)
      9.       └─rtweet:::.status_lookup(statuses[from:to], token = token)
     10.         └─rtweet:::TWIT(get = get, url, token)
     11.           └─httr::GET(url, ...)
     12.             └─httr:::request_perform(req, hu$handle$handle)
     13.               └─httr:::perform_callback("request", req = req)
     14.                 └─webmockr:::callback(...)
     15.                   └─webmockr::HttrAdapter$new()$handle_request(req)
     16.                     └─private$request_handler(req)$handle()
     17.                       └─eval(parse(text = req_type_fun))(self$request)
     18.                         └─err$run()
     19.                           └─self$construct_message()
    ── Error (test-pull_tweet_data.R:28:3): pull_tweet_data() is able to retrieve additional metadata starting with tweet IDs ──
    Error:

    ================================================================================
    An HTTP request has been made that vcr does not know how to handle:
    GET https://api.twitter.com/1.1/statuses/lookup.json?id=...................
    vcr is currently using the following cassette:
      - ../fixtures/metadata_from_ids.yml
        - record_mode: once
        - match_requests_on: method, uri
    Set `VCR_VERBOSE_ERRORS=TRUE` for more verbose errors
    If you're not sure what to do, open an issue https://github.com/ropensci/vcr/issues
    & see https://books.ropensci.org/http-testing
    ================================================================================

    Backtrace:
         █
      1. ├─vcr::use_cassette(...) test-pull_tweet_data.R:28:2
      2. │ └─cassette$call_block(...)
      3. └─tidytags::pull_tweet_data(id_vector = sample_tags$id_str, n = 10) test-pull_tweet_data.R:29:4
      4.   ├─base::ifelse(...)
      5.   ├─base::ifelse(...)
      6.   └─rtweet::lookup_statuses(id_vector[1:n])
      7.     ├─base::do.call("lookup_statuses_", args)
      8.     └─rtweet:::lookup_statuses_(...)
      9.       └─rtweet:::.status_lookup(statuses[from:to], token = token)
     10.         └─rtweet:::TWIT(get = get, url, token)
     11.           └─httr::GET(url, ...)
     12.             └─httr:::request_perform(req, hu$handle$handle)
     13.               └─httr:::perform_callback("request", req = req)
     14.                 └─webmockr:::callback(...)
     15.                   └─webmockr::HttrAdapter$new()$handle_request(req)
     16.                     └─private$request_handler(req)$handle()
     17.                       └─eval(parse(text = req_type_fun))(self$request)
     18.                         └─err$run()
     19.                           └─self$construct_message()
    ── Error (test-pull_tweet_data.R:49:3): pull_tweet_data() is able to retrieve additional metadata starting with tweet URLs ──
    Error:

    ================================================================================
    An HTTP request has been made that vcr does not know how to handle:
    GET https://api.twitter.com/1.1/statuses/lookup.json...............
    vcr is currently using the following cassette:
      - ../fixtures/metadata_from_urls.yml
        - record_mode: once
        - match_requests_on: method, uri
    Set `VCR_VERBOSE_ERRORS=TRUE` for more verbose errors
    If you're not sure what to do, open an issue https://github.com/ropensci/vcr/issues
    & see https://books.ropensci.org/http-testing
    ================================================================================

    Backtrace:
         █
      1. ├─vcr::use_cassette(...) test-pull_tweet_data.R:49:2
      2. │ └─cassette$call_block(...)
      3. └─tidytags::pull_tweet_data(...) test-pull_tweet_data.R:50:4
      4.   ├─base::ifelse(...)
      5.   └─rtweet::lookup_statuses(get_char_tweet_ids(url_vector[1:n], url_vector = url_vector[1:n]))
      6.     ├─base::do.call("lookup_statuses_", args)
      7.     └─rtweet:::lookup_statuses_(...)
      8.       └─rtweet:::.status_lookup(statuses[from:to], token = token)
      9.         └─rtweet:::TWIT(get = get, url, token)
     10.           └─httr::GET(url, ...)
     11.             └─httr:::request_perform(req, hu$handle$handle)
     12.               └─httr:::perform_callback("request", req = req)
     13.                 └─webmockr:::callback(...)
     14.                   └─webmockr::HttrAdapter$new()$handle_request(req)
     15.                     └─private$request_handler(req)$handle()
     16.                       └─eval(parse(text = req_type_fun))(self$request)
     17.                         └─err$run()
     18.                           └─self$construct_message()

    [ FAIL 7 | WARN 0 | SKIP 2 | PASS 62 ]
    Error: Test failures

### check `tidytags` for goodpractice:

In all your functions (.R files) and in the tests listed below, the package {goodpractice} detected long code lines (above 80 characters).

<details> <summary><strong>Test files with long code lines</strong></summary>

tidytags/tests/testthat/test-geocode\_tags.R
tidytags/tests/testthat/test-get\_char\_tweet\_ids.R
tidytags/tests/testthat/test-get\_url\_domain.R
tidytags/tests/testthat/test-lookup\_many\_tweets.R
tidytags/tests/testthat/test-pull\_tweet\_data.R

## Check package metadata files

### [README](https://github.com/bretsw/tidytags)

-   Instead of “Simple Collection and Powerful Analysis of Twitter Data”, I would write “Simple Collection and Powerful Analysis of Twitter Data collected **with TAGS**”.
[…]
