### 2021-09-23 #382 @bretsw
Hi @maelle! Thank you for checking in again. I think @jrosen48 and I are (finally) getting settled. We set a meeting on October 6 to start tackling the requested changes. Excited to get back to this.
