### 2021-11-23 #382 @bretsw
We're getting there! Fixed all the VCR and testing issues and have updated quite a bit of the requested explanations. Still a little bit to do, but we're making significant progress.
