### 2021-12-20 #382 @bretsw
@maelle, @llrs, and @marionlouveaux, thank you for you patience. Despite our silence, @jrosen48 and I have been hard at work this fall making the requested updates. We are ready for you to take a look, whenever it is convenient for you all (acknowledging that it has taken us a very long time on our end). Thank you for your comments and suggestions; {tidytags} is a much better package now because of them.

I'll leave our complete response in the next comment.
