### 2021-12-20 #382 @bretsw
# Reviewer 1 Comments

## VIGNETTES

1. The setup vignette is not clear what steps are necessary and which not, I would suggests adding titles and an index.

Thank you for this feedback. In response, we have noted that the first three key tasks (previously pain points) are required, while the fourth key task is optional and required only for geocoding. We are amenable to adding titles and an index; can you help us to understand what you mean by these? We think there is a title for the “Getting started with tidytags” vignette and there are also section headings for the vignette.

2. The vignettes have a "Pain point #4" which I couldn't find referenced anywhere. Also, perhaps using the more descriptive title would make it easier for users to know what it is about (Removing the Pain point reference entirely of the title of the section). Not that they aren't pain points but just redirect users when needed to the solutions/documentation as they go.

Thank you for pointing out this oversight. We mistakenly did not list all four of the key tasks (previously pain points), instead listing only three. We have corrected this in the vignette. Also, as noted here in and in our response to the above feedback, we have renamed the “pain points” key tasks to more accurately describe their nature.

However, most of the code chunks of the vignette are not run (as reported by BiocCheck):

Thank you for pointing this out. We acknowledge this but note that this is a function of the nature of this vignette on getting started with the package. We also note that there are extensive tests that ensure that the functionality of the package is sound, and so did not yet make any changes in response to this feedback.

3. And of those run are adding documentation or set up of the vignette. Perhaps some kind of setup specific for the vignettes could be used, otherwise they defeat their purpose and turn into plain READMEs. (I know it is not easy for CRAN, so maybe set them up as articles just on the website but outside CRAN?)

The vignette does compile on real data, it has just been precomputed following this guide: https://ropensci.org/blog/2019/12/08/precompute-vignettes/. This means that we do not have to wait for a lengthy compilation each time we rebuild the website.

4. To create the google API key step is not clear enough (perhaps a redesign on API configuration interface?). An indication to use Google Sheet API to the question "Find out what kind of credentials you need?", would be helpful.

Thank you for this feedback. Indeed, the process for obtaining a key for the Google Sheets API had changed. We have a) updated this section of the Getting Started vignette and b) added screenshots to facilitate the process of creating and storing (within R/RStudio) this key.

5. It should be pointed out that OpenCage Geocoding API key is not needed to use the package. Also the discussion about the price and API limits might be good for an issue but doesn't fit well on the vignette (I've seen that @maelle asked for this, but now that the package is settle in Open Cage maybe it is no longer needed or it can be reduced).

Thank you for pointing this out. As noted above, we have labeled this as optional and required only for geocoding.

6. On the chunk about "dplyr::glimpse(example_after_rtweet)" I get a different result 2204 rows compared to the 2,215 reported on the vignette.

Thank you for pointing this out. This is likely because some users have deleted their accounts or tweets in the period in between us creating this vignette and your access to the data. We have added the following note to the vignette:

“Note that this number of rows is at the point at which we collected the data; the number may differ when you carry out this same search.”

7. When I run the following code chunk I get an error (as I don't have the package longurl yet): `example_domains <- get_url_domain(example_urls)`

Thank you for pointing this out. We have added this note to the vignette: “Please note that these two packages are required to use this function (but they are not installed automatically when {tidytags} is installed.”

8. Before using a package in Suggests, it should be tested if they can be loaded (you can use `rlang::is_installed(longurl) or requireNamespace("longurl", quietly = TRUE)` ).

Thank you for this suggestion. We have made an addition above where the function requiring {longurl} is used to test whether it can be loaded (and to share with users how to install the package if it is not).

9. Last, I don't know how to push data to the google sheet TAGS created on the first vignette.

Thank you for pointing this out. We should have clarified that our use of Google Sheets is unidirectional: We use TAGS to access data, but process the data in R, but do not plan for users of the package to update the TAGS sheet with the processed data. We added a note to the vignette describing this and noting that users could push data to the Google Sheet or store it as a CSV (among other options).

## FUNCTION DOCUMENTATION

10. Name of the function repeated on the description on `add_users_data`, I think it is not needed.

Thank you for pointing this out; we have removed the name of the function from the description.

## EXAMPLES

11. Examples cannot be run without the authentication setup and there is no mention of this on the help pages. Perhaps a minor comment will remind users.

We have added a brief comment to the documentation for the following functions: read_tags(), pull_tweet_data(), lookup_many_tweets(), geocode_tags(), get_upstream_tweets(), add_users_data()

This is the comment we added:
#' @details This function requires authentication; please see
#'   \code{vignette("setup", package = "tidytags")}

## COMMUNITY GUIDELINES

12. There isn't any BugReports but on the vignettes there is info about how to get help. Would suggest to add the issues link on the description too. The contributing file is extensive and well organized.

Thank you for pointing this out. We have added a BugReports field to the DESCRIPTION.

## paper.md

### STATEMENT OF NEED

13. The authors say: "Yet, many approaches to collecting social media data in the moment require important technical skill that may dissuade social scientists from getting started". However this packages requires authentication in 2 or 3 API and a special google sheets. I have some doubts tidytags will attract the target audience.

Thank you for pointing this out. We have made several changes to address this comment. First, we have changed some of the language we use to justify the need for the package to not overstate the degree to which the package is accessible to novices (though, as we note next, we think the package is fundamentally accessible to relative newcomers to R). Second, we have added extensive documentation-many of these additions in response to reviewer feedback-especially in the ‘Getting started with tidytags’ vignette as well as in the paper accompanying the package. Lastly, while we have made the two aforementioned changes and thank the reviewers for these suggestions that have improved our package, we note that we still think the package can be used by those relatively new to R in a way that is easier than other extant approaches for accessing Twitter data.

### REFERENCES

14. None of the references have dois or urls to online resources.

We have added DOI URLs for each reference in our .bib file.

15. There's an additional `"` on the yaml heading of paper.md that prevented viewing the paper.

Thank you for catching this typo. We have corrected it.

## TESTS

16. I got 2 tests that failed and 3 warnings (besides 2 that were skipped). The test-get_url_domain.R:14:5 test reported `domain4` not equal to "npr.org", on the browser I get asked for cookie consent on the browser, when run locally, outside testthat or vcr, I get the url of choice.npr.org.

This a persistent bug in {vcr} that turned out to be tricky to resolve (see https://github.com/ropensci/vcr/issues/220). We have removed the two skipped tests that depended on this.

17. The other failing test are weird (as I don't get them when I run them on the R console but only on the build/check Rstudio panel).

We have updated all tests, and all are currently passing locally and in GitHub CI.

18. I have a development version installed of vcr and one of the warnings is related to it. The new version warns when the cassettes are empty, this in my experience means that the test it not conclusive, but this could also be related to not having the geo code API enabled.

We have not seen an error like this on our end.

19. The other warnings are on test-get_url_domain.R, lines 3 and 32, Invalid URL I'm not sure why, because when I paste on my browser I get redirected to https://www.aect.org/about_us.php. (BTW perhaps the link can be changed to https instead of http).

This is specifically related to the error in 1.16, a persistent bug in {vcr} that turned out to be tricky to resolve (see https://github.com/ropensci/vcr/issues/220). We have removed the two skipped tests that depended on this.

## ADD REVIEWER

20. Should the author(s) deem it appropriate, I agree to be acknowledged as a package reviewer ("rev" role) in the package DESCRIPTION file.

We have added name and ORCID to DESCRIPTION file (https://github.com/llrs).

## REVIEW COMMENTS

21. The package rtweet is experimenting drastic changes (I'm involved on rtweet maintenance) and there will be a major release with breaking changes. Probably it will break this package (the recommendation about the token will change for instance), so be ready to update it accordingly.

At the time of our revisions (November 2021), {rtweet} has not yet been updated; it is still version 0.7.0.

22. The package contains relative few simple functions that provide more data from twitter or make it easier to analyze it. I have not analyzed Twitter data (except for an overview of a user account), so don't know how useful the data is. I am not a user of TAGs but I'm a bit puzzles how to add information to the google sheet: if I'm a new user how should I do it? I mean I can get the template but how I fill it? I think this package would be easier for non technical people if it included a function to add the information gathered via rtweet or processed with the package back to the original google sheet.

Please see our response to comment 1.9.

23. Haven't fully read the paper.md for JOSS but I think it is short enough and comprehensive of the package.

From a more technical point of view, I have some comments about the code and the package:

24. There are 75 lines longer than 80 characters, try to reduce them. Probably it is just a matter of style and perhaps creating new shorter variables

Thank you for catching this. We have corrected all lines longer than 80 characters and checked with goodpractice::goodpractice().

25. Also namespaces in Imports field not imported from: ?gargle? ?readr?. All declared Imports should be used.

We have removed gargle from the Imports list. The gargle package is mentioned in the “Getting started with tidytags” vignette, but the package is not called anywhere in tidytags. However, because readr is used extensively throughout the testthat scripts, so we have moved readr to the Suggests list.

26. The `get_char_tweet_ids` function could be improved, with only one argument if it is a data.frame then extract the status_id and get the ID via id_str. If it is an url you can just extract the last numbers with `gsub("https?\\://twitter.com\\/.+/statuses/", "", df$status_url)`, no need to modify the data.frame and then extract the vector again.

Thank you for this suggestion. We have updated the get_char_tweet_ids() function and shortened the code significantly.

27. On `process_tweets` you can simplify the is_self_reply to `ifelse(.data$is_reply & .data$user_id == .data$reply_to_user_id, TRUE, FALSE)`.

Thank you for this suggestion. We have updated the function accordingly.

28. On `get_upstream_replies` the examples are not informative, as there are no replies to get data from on the example dataset. You make multiple calls to `pull_tweet_data`, some of them might be unnecessary. The `process_tweets` can be called just once at the end instead of multiple times and on each loop run. This should speed up the process. Also, if there are at most 90000 tweets taken from each run, then you can estimate the number of iterations needed and inform the user. This might make the wait easier. Perhaps it would be better to use `lookup_many tweets` as it does a similar process. However, users might hit the rate limit and I don't see any information being passed to the user regarding this.

Thank you for pointing out this oversight. We have updated the example for `get_upstream_replies()` to include replies in the example dataset. We also updated the name of this function to `get_upstream_tweets()` to more accurately describe the returned data (i.e., not all the tweets retrieved from “upstream” the reply thread will also be replies; some will just be regular tweets).

We also changed the function so that `process_tweets()` is not called at all.

To make the code more readable, we created an internal function `flag_unknown_upstream()` and added tests.

Finally, we added a note in the function documentation that it is impossible to know in advance how far “upstream” you can trace back a reply thread, and that running `get_upstream_tweets()` might take a while.

29. Looking at `create_edgelist`, it calls `process_tweets` and also `get_replies`, `get_retweets`, `get_quotes`, `get_mentions` which they call `process_tweets` too. Perhaps some internal functions could be created to avoid calling process_tweets multiple times on the same data.

We have changed the names of these functions so that they do not mask functions imported from {rtweet}. For instance, `get_replies()` is now `create_replies_edgelist()`.

Rather than creating an internal function, we simply removed the redundant `process_tweets()` that occurred within `create_edgelist()`. We do need to keep `process_tweets()` within the `create_` functions because they can serve as standalone functions, not just internal functions for the more expansive `create_edgelist()`.

In addition, we have updated `create_edgelist()` altogether to accept an input parameter of `type`. See comments 2.9 and 2.47 for more details.

---

# Reviewer 2 Comments

## README

1. In Overview of the README, I would add a sentence to explain what is TAGS, and explain a bit more in detail how rtweet and opencage are used in tidytags and what else is provided by tidytags.

Thank you for pointing this out. We have made several revisions to the Overview section of the README to better introduce the roles of TAGS, rtweet, and opencage. The section now reads as follows:

The purpose of {tidytags} is to make the collection of Twitter data more accessible and robust. {tidytags} retrieves tweet data collected by a [Twitter Archiving Google Sheet (TAGS)](https://tags.hawksey.info/), gets additional metadata from Twitter via the [{rtweet}](https://docs.ropensci.org/rtweet/index.html) package, and from OpenCage using the [opencage](https://opencagedata.com/) package, and provides additional functions to facilitate a systematic yet flexible analyses of data from Twitter. TAGS is based on Google spreadsheets. A TAGS tracker continuously collects tweets from Twitter, based on predefined search criteria and collection frequency.

In short, {tidytags} first uses TAGS to easily collect tweet ID numbers and then uses the R package {rtweet} to re-query the Twitter API to collect additional metadata.

{tidytags} also introduces functions developed to facilitate systematic yet flexible analyses of data from Twitter. It also interfaces with several other packages, including the [opencage package](https://opencagedata.com/), to geocode the locations of Twitter users based on their biographies.

Two vignettes illustrate the setup and use of the package:

- [Getting started with tidytags](https://bretsw.github.io/tidytags/articles/setup.html)
- [Using tidytags with a conference hashtag](https://bretsw.github.io/tidytags/articles/tidytags-with-conf-hashtags.html)

## EXAMPLES

2. Except for lookup_many_tweets

Thank you for catching this; we have added examples to lookup_many_tweets(). Note that we have not included an example of a dataset of more than 90,000 tweets because this would be prohibitively large.

## COMMUNITY GUIDELINES

3. Missing `BugReports`

Thank you, we have added this to the DESCRIPTION.

## FUNCTIONALITY

4. 7 tests out of 69 failed.

See our response to comments 1.16-1.19. We have updated all tests and found that all are passing for us locally and in GitHub CI.

## ADD REVIEWER

5. Should the author(s) deem it appropriate, I agree to be acknowledged as a package reviewer ("rev" role) in the package DESCRIPTION file.

We have added name and ORCID to DESCRIPTION file (https://github.com/marionlouveaux).

## REVIEW COMMENTS

6. The {tidytags} package gives the possibility to read a TAGS tracker, which is a Google app continuously collecting tweets from Twitter, based on predefined search criteria and collection frequency. It provides wrappers to {rtweet} and {opencage} functions to simplify the retrieval of metadata, either not fetched by TAGS or not existing in Twitter (in the case of geocoding). In addition, it provides functionalities to compute additional descriptive variables about the collected tweets and to visualise relationships between tweets. The {tidytags} package interacts with 3 APIs (Google spreadsheets, Twitter and OpenCage) and one Google app (TAGS). For this reason, the set up is a bit long and tedious when done from scratch. The package itself contains a small number of functions, that are well documented.

Thank you for this feedback. We received similar feedback about the setup from the other reviewer, as well, and have made some changes to the setup, namely, emphasizing what is strictly necessary to do and what is optional for particular functionality (see Reviewer 1, Comment and Response 1). Also, as noted in a response to a comment below (18), we have added a checklist to facilitate the process of getting started with the package.

7. I used the {pkgreviewr} from RopenSci to conduct my review (a big thanks to the authors of this package). I configured TAGS and created a Gooogle API key. I already had the configuration for {rtweet} and {opencage}. I could run {tidytags} on my own TAGS tracker (and it worked!).

### My main comments concern:

8. the description of the package goals. I didn't get a clear idea of what {tidytags} is doing at the first glimpse on the README. Some more sentences in the Overview paragraph would help. In addition, I would recommend adding a sketch with the main functions and the link with the different APIs. I would also add a checklist of all the things you need to have in order to be set up (more precise than the 4 pain points).

Thank you for pointing this out. We have added information from the paper to the Overview section of the README (as also noted in response to comment 1).

9. simplification of code for the functions get_replies, get_quotes, get_retweets and get_mentions. I would create an internal get_content function that takes as input parameters df and type, type being "reply", "quote", "retweet" or "mentions".

We have removed these four specialized functions. Instead, we created a new function `filter_by_tweet_type()` with input parameter `type` that can be set to reply, retweet, quote, or mention. We then updated `create_edgelist()` to call `filter_by_tweet_type()`, meaning that `create_edgelist()` accepts an input parameter `type` that can be set to reply, retweet, quote, or mention, but `type` defaults to “all”.

[…]
