### 2021-12-21 #382 @llrs
Hi @bretsw I've read the response but I probably won't have time to look into the package and replies with detail until a couple of weeks from now.

@maelle Can we use the `@ropensci-review-bot` bot now on this issue? Just to check the package online and not only on my machine (I might forget some of the checks)
