### 2021-12-21 #382 @bretsw
I've removed the .DS_Store files (and added this to .gitignore).
I've fixed the lines longer than 80 characters.
I've added the 'codemeta.json' file.

Should be ready for @ropensci-review-bot check package again.
