### 2021-12-27 #382 @llrs
I'm testing it now with R 4.1.2

- 1.1
 There a missing `[` on the first line of the [Key task 1](https://bretsw.github.io/tidytags/articles/setup.html#key-task-1-making-sure-your-tags-tracker-can-be-accessed). I'm not sure now what I meant but I find the titles and indices of the vignettes are now appropriate Note that there are some links related to the other viggnette that are hardcoded. It is usually recommended to look up `vignette("setup", package = "tidytags")`.
 Also note that users don't see the options used on the code chunks so mentions to " (notice that eval = FALSE)." (Found on the last text line before [pull_tweet_data](https://bretsw.github.io/tidytags/articles/tidytags-with-conf-hashtags.html#pull-tweet-data-))  are not useful for the reader without accessing the source of the vignette.

- 1.2
 The vignettes are now very clear
- 1.3
 As you wish
- 1.4
  Much better documented, thanks.
- 1.5
  Thanks, was the discussion about price keep on purpose?
- 1.6
  Great!
- 1.7
  `get_url_domain` uses longurl package without checking if it is installed, either add it to imports or check via `if (requireNamespace("longurl", quiet = TRUE) ) {...}`. `geocode_tags` uses unconditional opencage as you said it is optional I would use `if (requireNamespace("longurl", quiet = TRUE) ) {...}`. Check that other suggested packages are used conditionally (I think mapview is also used on the vignette but not checked if installed). So that the users receive a friendly message when using the function without the package (not only on the vignette [where I would delete this information])
- 1.8
  I couldn't find any `requireNamespace` on the package.
- 1.9
  Thanks, that will help the users.
- 1.13
  I'm sure the changes made will help all R users using the package.
- 1.21
  This comment was meant as a heads up of the work that will mean to maintain a package that depends on 3 different APIs an the upcomming changes on rtweet, not that the changes on rtweet could affect the review.
- 1.28
 If users have not set up correctly the rtweet account, they don't receive any information and the `tmp_df` from  `get_upstream_tweets` example is empty, resulting on an error when using `get_upstream_tweets`. Maybe adding a check on the function to see if some tweets were collected might be helpful.

Many thanks for all the changes.

Hours spend: ~3 (Is still needed to report this?)
