### 2022-01-03 #382 @maelle
Thanks @bretsw for the detailed response & @llrs for the new feedback!

Regarding empty cassettes, it might be worth subscribing to https://github.com/ropensci/vcr/issues/244
