### 2022-01-03 #382 @bretsw
@llrs, thank you for your follow-up review and comments. I have made further updates as requested:

- 1.1: I have corrected the missing "[", referenced vignettes everywhere using `vignette()`, and removed the comment about setting "eval = FALSE".
- 1.5: We accidentally missed deleting the discussion of price. I have now deleted this paragraph.
- 1.7 and 1.8: I have added `if (!requireNamespace("package_name", quietly = TRUE)) {stop("Please install the {package_name} package to use this function", call. = FALSE)}` for {longurl} and {urltools} inside the `get_url_domain()` function, for {opencage} inside the `geocode_tags()` function., and for {beepr} inside the `lookup_many_tweets()` and `lookup_many_users()` functions. I have added `if (requireNamespace("package_name", quietly = TRUE) for {mapview}, {sf}, {ggplot2}, {ggraph}, and {tidygraph} in the function documentation and vignettes. This should cover all suggested-but-not-imported packages.
- 1.28: I added `if (nrow(unknown_upstream) == 0) {stop("There are no upstream replies to get.", call. = FALSE)}` to the `get_upstream_tweets()` function. Please let me know if this is sufficient for the concern you have raised, or if there needs to be another check added to `get_upstream_tweets()` or even `pull_tweet_data()`.
