### 2022-01-03 #382 @bretsw
@maelle, could we remove the `holding` label for this thread?
