### 2022-01-03 #382 @bretsw
> I think that the package is now in great shape @bretsw, thanks for adding me as reviewer on your package description file.

Thank you, @llrs, for your feedback, guidance, and patience. I have been very grateful for this review process.
