### 2022-01-04 #382 @maelle
@llrs thank you! Here's the reviewer approval comment template if you have time: https://devguide.ropensci.org/approval2template.html

@marionlouveaux would you have time to take a last look at the package?
