### 2022-01-29 #382 @marionlouveaux
Hi all,

Thank you for responding to my review and adding me as a reviewer on your package.

In the help of pull_tweet_data(), I noticed some formatting issues such as \code{n} (in the example section). I suggest using roxygen2md::roxygen2md() (https://roxygen2md.r-lib.org/).
