### 2022-01-29 #382 @marionlouveaux
## Reviewer Response

Thank you for responding to my review and adding me as a reviewer on your package. The package documentation is much better and much clearer. I found one minor aesthetic issue in the help of `pull_tweet_data()`: I noticed some formatting issues such as `\code{n}` (in the example section). I suggest using `roxygen2md::roxygen2md()` to fix these (see https://roxygen2md.r-lib.org/).

I still approve the package now so as not to delay further the process.

#### Final approval (post-review)

Estimated hours spent reviewing: 1
