### 2022-01-29 #382 @bretsw
@marionlouveaux, thank you so much for your follow up review. @jrosen48 and I are truly grateful!

Thank you for pointing out `roxygen2md::roxygen2md()`--what a handy function! I've cleaned up the documentation and pushed changes.
