### 2022-01-31 #382 @ropensci-review-bot
Approved! Thanks @bretsw for submitting and @llrs, @marionlouveaux for your reviews! :grin:
