### 2022-01-31 #382 @bretsw
Hi @maelle, so I thought I updated everything as requested for the transfer to rOpenSci, but now all my GitHub actions are failing. Any suggestions? https://github.com/ropensci/tidytags/actions
