### 2022-01-31 #382 @bretsw
>

Looks like I fixed the GitHub pages issue (needed to change the build to happening from the root directory in Settings) and got codecov to work (needed to update the key following the transfer to rOpenSci) - https://github.com/ropensci/tidytags

For the R-CMD-CHECK, looks like some issue with the terra package is tripping it up. So, probably nothing to be done at this point.
